Move both check and check-external into same command module, add ignore directive handling to check-external
