+++ src/commands/check/diagnostics.rs
