+++ src/commands/mod.rs
@@ -1,5 +1,4 @@
-pub mod check_external;
-pub mod check_internal;
+pub mod check;
 pub mod report;
 pub mod server;
 pub mod sync;
